test threads can be deleted by authorized users only

+++ app/Http/Controllers/ThreadsController.php
@@ -114,6 +114,14 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        if($thread->user_id != auth()->id()) {
+            if(\request()->wantsJson()) {
+                return response(['status' => 'Permission denied'], 403);
+            }
+
+            return redirect('/login');
+        }
+
         $thread->replies()->delete();
         $thread->delete();
         if(\request()->wantsJson()) {
